+++ MainWindow.xaml.cs
@@ -191,6 +191,8 @@ public partial class MainWindow : Window
 
         ContentBox.WordWrap = conf.WordWrap;
         ContentBox.ShowLineNumbers = conf.ShowLineNumbers;
+        ContentBox.TextArea.Options.ShowSpaces = conf.ShowSpaces;
+        ContentBox.TextArea.Options.ShowEndOfLine = conf.ShowLineFeed || conf.ShowCarriageReturn;
         ContentBox.HorizontalScrollBarVisibility = conf.WordWrap
             ? System.Windows.Controls.ScrollBarVisibility.Disabled
             : System.Windows.Controls.ScrollBarVisibility.Auto;